refactor(mvt): remove ForegroundService, manage secondary display via Activity lifecycle
Remove ForegroundService to eliminate Google Play policy concerns and simplify
the implementation. Secondary display presentation is now managed directly in
MainActivity's onResume/onPause lifecycle:

- Remove FOREGROUND_SERVICE and FOREGROUND_SERVICE_DATA_SYNC permissions
- Remove TerminalService class and service declaration from manifest
- Move DisplayListener and Presentation management to MainActivity
- Dismiss presentation when app goes to background (onPause)
- Restore presentation when app returns to foreground (onResume)

This change means secondary display content is only shown while the app is
in the foreground, but eliminates foreground service policy requirements.
