+++ mobile-vibe-terminal/src/androidMain/AndroidManifest.xml
@@ -7,10 +7,6 @@
     <uses-permission android:name="android.permission.REQUEST_INSTALL_PACKAGES" />
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
 
-    <!-- Permissions for secondary display -->
-    <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
-    <uses-permission android:name="android.permission.FOREGROUND_SERVICE_DATA_SYNC" />
-
     <application
         android:name=".VibeTerminalApp"
         android:icon="@mipmap/ic_launcher"
@@ -30,14 +26,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <!-- Foreground Service for secondary display -->
-        <service
-            android:name=".service.TerminalService"
-            android:enabled="true"
-            android:exported="false"
-            android:foregroundServiceType="dataSync" />
-
 
         <!-- FileProvider for sharing APK files -->
         <provider
