+++ mobile-vibe-terminal/src/androidMain/kotlin/tokyo/isseikuzumaki/vibeterminal/MainActivity.kt
@@ -1,12 +1,22 @@
 package tokyo.isseikuzumaki.vibeterminal
 
+import android.content.Context
 import android.content.res.Configuration
+import android.hardware.display.DisplayManager
 import android.os.Bundle
+import android.view.Display
 import android.view.KeyEvent
+import android.view.WindowManager
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import cafe.adriel.voyager.navigator.Navigator
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.GlobalContext
 import org.koin.core.context.startKoin
@@ -14,7 +24,10 @@ import tokyo.isseikuzumaki.vibeterminal.di.appModule
 import tokyo.isseikuzumaki.vibeterminal.di.dataModule
 import tokyo.isseikuzumaki.vibeterminal.di.platformModule
 import tokyo.isseikuzumaki.vibeterminal.input.HardwareKeyboardHandler
-import tokyo.isseikuzumaki.vibeterminal.platform.launchSecondaryDisplay
+import tokyo.isseikuzumaki.vibeterminal.service.TerminalPresentation
+import tokyo.isseikuzumaki.vibeterminal.service.isValidSecondaryDisplay
+import tokyo.isseikuzumaki.vibeterminal.terminal.DisplayTarget
+import tokyo.isseikuzumaki.vibeterminal.terminal.TerminalDisplayManager
 import tokyo.isseikuzumaki.vibeterminal.terminal.TerminalStateProvider
 import tokyo.isseikuzumaki.vibeterminal.ui.components.TriggerEventHost
 import tokyo.isseikuzumaki.vibeterminal.ui.screens.ConnectionListScreen
@@ -22,6 +35,30 @@ import tokyo.isseikuzumaki.vibeterminal.ui.theme.VibeTerminalTheme
 import tokyo.isseikuzumaki.vibeterminal.util.Logger
 
 class MainActivity : ComponentActivity() {
+
+    private var presentation: TerminalPresentation? = null
+    private lateinit var displayManager: DisplayManager
+    private val activityScope = CoroutineScope(SupervisorJob() + Dispatchers.Main)
+    private var isResumed = false
+
+    // Display listener for monitoring secondary display connection/disconnection
+    private val displayListener = object : DisplayManager.DisplayListener {
+        override fun onDisplayAdded(displayId: Int) {
+            Logger.d("Display added: $displayId")
+            updateConnectionState()
+        }
+
+        override fun onDisplayRemoved(displayId: Int) {
+            Logger.d("Display removed: $displayId")
+            updateConnectionState()
+        }
+
+        override fun onDisplayChanged(displayId: Int) {
+            Logger.d("Display changed: $displayId")
+            updateConnectionState()
+        }
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -39,16 +76,24 @@ class MainActivity : ComponentActivity() {
             }
         }
 
-        // Start TerminalService to monitor secondary display connections
-        try {
-            launchSecondaryDisplay(this)
-            Logger.d("MainActivity: TerminalService started for display monitoring")
-        } catch (e: Exception) {
-            Logger.e(e, "MainActivity: Failed to start TerminalService")
-        }
+        // Initialize DisplayManager
+        displayManager = getSystemService(Context.DISPLAY_SERVICE) as DisplayManager
 
         // Check initial hardware keyboard state
         updateHardwareKeyboardState(resources.configuration)
+
+        // Observe terminalDisplayTarget to control presentation
+        activityScope.launch {
+            TerminalDisplayManager.terminalDisplayTarget.collect { target ->
+                Logger.d("MainActivity: terminalDisplayTarget changed to $target")
+                if (isResumed) {
+                    when (target) {
+                        DisplayTarget.SECONDARY -> updatePresentation()
+                        DisplayTarget.MAIN -> dismissPresentation()
+                    }
+                }
+            }
+        }
 
         enableEdgeToEdge()
         setContent {
@@ -58,6 +103,116 @@ class MainActivity : ComponentActivity() {
                 }
             }
         }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        isResumed = true
+        Logger.d("MainActivity: onResume - registering display listener")
+
+        // Register display listener
+        displayManager.registerDisplayListener(displayListener, null)
+
+        // Check for existing valid secondary displays
+        updateConnectionState()
+
+        // Show presentation if target is secondary
+        if (TerminalDisplayManager.terminalDisplayTarget.value == DisplayTarget.SECONDARY) {
+            updatePresentation()
+        }
+    }
+
+    override fun onPause() {
+        super.onPause()
+        isResumed = false
+        Logger.d("MainActivity: onPause - unregistering display listener and dismissing presentation")
+
+        // Dismiss presentation when app goes to background
+        dismissPresentation()
+
+        // Unregister display listener
+        displayManager.unregisterDisplayListener(displayListener)
+
+        // Reset display connection state
+        TerminalDisplayManager.setDisplayConnected(false)
+        TerminalStateProvider.clearSecondaryDisplayMetrics()
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        Logger.d("MainActivity: onDestroy")
+        activityScope.cancel()
+    }
+
+    private fun updateConnectionState() {
+        val hasValidSecondary = displayManager.displays.any { it.isValidSecondaryDisplay() }
+        Logger.d("updateConnectionState: hasValidSecondary=$hasValidSecondary")
+        TerminalDisplayManager.setDisplayConnected(hasValidSecondary)
+
+        if (!hasValidSecondary) {
+            TerminalStateProvider.clearSecondaryDisplayMetrics()
+        }
+    }
+
+    private fun updatePresentation() {
+        if (!isResumed) {
+            Logger.d("updatePresentation: Activity not resumed, skipping")
+            return
+        }
+
+        if (presentation != null) {
+            Logger.d("Presentation already exists")
+            return
+        }
+
+        val displays = displayManager.displays
+        Logger.d("Found ${displays.size} displays")
+
+        val secondaryDisplay = displays.firstOrNull { it.isValidSecondaryDisplay() }
+
+        if (secondaryDisplay != null) {
+            Logger.d("Found valid secondary display: ${secondaryDisplay.displayId}, flags=${secondaryDisplay.flags}")
+            showPresentation(secondaryDisplay)
+        } else {
+            Logger.w("No valid secondary display found")
+        }
+    }
+
+    private fun showPresentation(display: Display) {
+        try {
+            Logger.d("Creating presentation for display ${display.displayId}")
+
+            presentation = TerminalPresentation(
+                outerContext = this,
+                display = display,
+                onDisplaySizeCalculated = { cols, rows, widthPx, heightPx ->
+                    Logger.d("Secondary display size calculated: ${cols}x${rows} (${widthPx}x${heightPx}px)")
+                    TerminalStateProvider.requestResize(cols, rows, widthPx, heightPx)
+                }
+            )
+
+            presentation?.show()
+            Logger.d("Presentation shown successfully")
+
+            // Request focus restoration to main display after a short delay
+            activityScope.launch {
+                delay(200)
+                Logger.d("Requesting focus restoration to main display")
+                TerminalDisplayManager.requestFocusRestoration()
+            }
+        } catch (e: WindowManager.InvalidDisplayException) {
+            Logger.e(e, "InvalidDisplayException when showing presentation")
+            presentation = null
+        } catch (e: Exception) {
+            Logger.e(e, "Exception in showPresentation")
+            presentation = null
+        }
+    }
+
+    private fun dismissPresentation() {
+        Logger.d("Dismissing presentation")
+        presentation?.dismiss()
+        presentation = null
     }
 
     override fun dispatchKeyEvent(event: KeyEvent): Boolean {
